pyproject.toml
[project]
name = "wunderunner"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "docker>=7.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pydantic-ai>=0.0.40",
    "pydantic-graph>=0.0.40",
    "pydantic-settings>=2.5.0",
    "typer>=0.12.0",
    "rich>=13.8.0",
]

[project.scripts]
wunderunner = "wunderunner.cli.main:app"
wxr = "wunderunner.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/wunderunner"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"
